+++ tests/test_result/test_result_map.py
@@ -8,16 +8,16 @@ def test_map_success():
     assert Success(5).map(str) == Success('5')
 
 
-def test_map_failure():
+def test_alt():
     """Ensures that Failure is mappable."""
     assert Failure(5).map(str) == Failure(5)
-    assert Failure(5).map_failure(str) == Failure('5')
+    assert Failure(5).alt(str) == Failure('5')
 
 
 def test_fix_success():
     """Ensures that Success.fix is NoOp."""
     assert Success(5).fix(str) == Success(5)
-    assert Success(5).map_failure(str) == Success(5)
+    assert Success(5).alt(str) == Success(5)
 
 
 def test_fix_failure():